#8 Refactor rest-blog and add many fixes

+++ blog/rest-blog/src/main/java/com/github/srcmaxim/restblog/PostService.java
@@ -5,7 +5,9 @@ import io.quarkus.panache.common.Sort;
 import javax.enterprise.context.ApplicationScoped;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static javax.transaction.Transactional.TxType.REQUIRED;
 import static javax.transaction.Transactional.TxType.SUPPORTS;
@@ -22,16 +24,22 @@ public class PostService {
 
     @Transactional(SUPPORTS)
     public Post findPostById(String id) {
-        return Post.find("select p from Post p join fetch p.tags where p.id = ?1", id).firstResult();
+        return Post.findById(id);
     }
 
     public Post persistPost(@Valid Post post) {
+        post.id = Arrays.stream(post.title.split("\\s"))
+                .map(String::toLowerCase)
+                .collect(Collectors.joining("-"));
         Post.persist(post);
         return post;
     }
 
     public Post updatePost(@Valid Post post) {
         Post entity = Post.findById(post.id);
+        if (entity == null) {
+            return null;
+        }
         entity.title = post.title;
         entity.content = post.content;
         entity.readMinutes = post.readMinutes;
@@ -42,6 +50,8 @@ public class PostService {
 
     public void deletePost(String id) {
         Post post = Post.findById(id);
-        post.delete();
+        if (post != null) {
+            post.delete();
+        }
     }
 }
